Move: Pull Worker::Script::Source and friends into a separate header.
All of `WorkerSource` except its last few lines are a pure cut-and-paste from Worker::Script.

This is needed so that we can use `WorkerSource` in io-channels.h to define a dynamic isolate channel. io-channels.h cannot include worker.h as this would be a cyclic dependency.
